requirements.txt
requests==2.21.0
urllib3==1.24.2
boto3==1.12.2
botocore==1.15.2
numpy==1.18.1
pandas==1.0.1
python-dotenv==0.14.0